Improve AI document analysis accuracy for Brazilian invoices
Refine DANFE parsing instructions, data mapping, and confidence score handling to enhance AI read accuracy.

+++ server/ai-validation-schema.ts
@@ -13,7 +13,7 @@ export const aiAnalysisResponseSchema = z.object({
   documento: z.string().optional(), // CNPJ/CPF
   cliente_fornecedor: z.string().optional(),
   observacoes: z.string().optional(),
-  confidence: z.number().int().min(0).max(100, "Confiança deve estar entre 0 e 100"),
+  confidence: z.number().min(0).max(100, "Confiança deve estar entre 0 e 100"),
 });
 
 export type AiAnalysisResponse = z.infer<typeof aiAnalysisResponseSchema>;
